app/requirements.txt
alembic==1.3.2
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
Fiona==1.8.13
Flask==1.1.1
Flask-API==2.0
Flask-Bcrypt==0.7.1
flask-blueprint==1.3.0
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.8.0
GeoAlchemy2==0.6.3
geopandas==0.6.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
Mako==1.1.0
MarkupSafe==1.1.1
munch==2.5.0
mysqlclient==1.4.6
numpy==1.18.1
pandas==1.0.0
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
pymongo==3.10.1
PyMySQL==0.9.3
pyproj==2.4.2.post1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3
requests==2.22.0
Rtree==0.9.4
scikit-learn==0.22.2.post1
scipy==1.4.1
Shapely==1.6.4.post2
six==1.13.0
sklearn==0.0
SQLAlchemy==1.3.12
urllib3==1.25.7
Werkzeug==0.16.0
python-osc==1.7.4
SoundFile==0.10.3.post1
pydub==0.24.1
pyOSC3==1.2